refactor party class to be global

+++ DunGen/Models/MobFactory.swift
@@ -49,8 +49,8 @@ class MobFactory {
                 break
             }
             
-            if let cr = Float(mob[iCR]) {
-                monsterList.append(Monster(name: mob[iName], armorClass: Int(mob[iAC])!, hitPoints: 0, initiativeBonus: Int(mob[iInitiative])!, image: "Avatar4", hitDice: mob[iHitDice], challengeRating: cr)) //Float(mob[iCR])))
+            if let cr = Double(mob[iCR]) {
+                monsterList.append(Monster(name: mob[iName], armorClass: Int(mob[iAC])!, hitPoints: 0, initiativeBonus: Int(mob[iInitiative])!, image: "Avatar4", hitDice: mob[iHitDice], challengeRating: cr)) //Double(mob[iCR])))
             }
             //init(name: String, armorClass: Int, hitPoints: Int, initiativeBonus: Int, image: String, hitDice: String, challengeRating: Double)
